Modfy Driver Business Layer

+++ DVLD_BusinessLayer/clsLicense.cs
@@ -197,7 +197,7 @@ namespace DVlD_BusinessLayer
         public static bool AlreadyHaveLicense(int PersonID,int LicenseClass)
         {
            clsDriver Driver= clsDriver.FindByPersonID(PersonID);
-            int DriverID = (Driver!=null)? Driver.DriverID:-1;
+            int DriverID = (Driver!=null)? (int)Driver.DriverID:-1;
             if (DriverID == -1)
                 return false;
             else
